+++ src/Products/PloneMeeting/tests/PloneMeetingTestCase.py
@@ -319,10 +319,10 @@ class PloneMeetingTestCase(unittest.TestCase, PloneMeetingTestingHelpers):
             attrs.update({'date': datetime.now()})
         if objectType == 'MeetingItem':
             if 'proposingGroup' not in attrs.keys():
-                cleanRamCacheFor('Products.PloneMeeting.ToolPloneMeeting.get_orgs_for_user')
-                proposingGroup = self.tool.get_orgs_for_user(suffixes=['creators'])
-                if len(proposingGroup):
-                    attrs.update({'proposingGroup': proposingGroup[0].UID()})
+                cleanRamCacheFor('Products.PloneMeeting.ToolPloneMeeting.get_org_uids_for_user')
+                proposingGroupUids = self.tool.get_orgs_for_user(suffixes=['creators'])
+                if len(proposingGroupUids):
+                    attrs.update({'proposingGroup': proposingGroupUids[0]})
         obj = getattr(folder, folder.invokeFactory(contentType, **attrs))
         if objectType == 'Meeting':
             self.setCurrentMeeting(obj)