+++ CHANGES.rst
@@ -9,11 +9,22 @@ Changelog
   [gbastien]
 - Fixed `monkey.validate` (load `monkey` in tests so it is taken into account).
   [gbastien]
-- Fixed `UnicodeDecodeError` in `CategoriesOfOtherMCsVocabulary` when an disabled
-  category was in a `MeetingConfig` ihaving special characters in it's title.
+- Fixed `UnicodeDecodeError` in `CategoriesOfOtherMCsVocabulary` when a disabled
+  category was in a `MeetingConfig` having special characters in it's title.
   [gbastien]
 - Do not fail in `PMGenerablePODTemplatesAdapter.get_all_pod_templates` when
   `portal_ploneMeeting` is not available (for example when testing `imio.pm.wsclient`).
+  [gbastien]
+- Optimized `ram.cache` configuration:
+  - Monkeypatched `zope.ramcache.Storage.getEntry` to update timestamp while
+    getting an existing entry;
+  - Adapted ToolPloneMeeting.get_orgs_for_user to no more return objects as
+    it uses `ram.cache`, parameter `the_objects=False` by default now;
+  - Adapted `global_cache` settings, set `maxEntries=100000`, `maxAge=86400`,
+    `cleanupInterval=3600` so cache is kept for a long time.
+  - Do not more `ram.cache` `Meeting.query_state` and `MeetingItem.query_state`,
+    performance test shows it is not necessary.
+
   [gbastien]
 
 4.2b19 (2021-11-08)